#12 completed mock data and refactor system design
