+++ src/main/java/com/example/splitly/application/serviceInterface/InterfaceUserInfo.java
@@ -1,0 +1,5 @@
+package com.example.splitly.application.serviceInterface;
+
+public interface InterfaceUserInfo {
+
+}
