+++ src/main/java/com/example/splitly/application/serviceInterface/explain.txt
@@ -1,0 +1,1 @@
+Define method handle for service
