+++ src/main/java/com/example/splitly/application/usecase/explain.txt
@@ -1,0 +1,1 @@
+Handle request from controller and call service
