+++ src/main/java/com/example/splitly/domain/constants/explain.txt
@@ -1,0 +1,1 @@
+Define const variable in here!
