+++ src/main/java/com/example/splitly/domain/enumerator/explain.txt
@@ -1,0 +1,1 @@
+Define enum type in here!
