+++ src/main/java/com/example/splitly/infrastructure/data/migration/explain.txt
@@ -1,0 +1,1 @@
+migration data in database if db change table, field,..
