+++ src/main/java/com/example/splitly/presentation/dto/response/ResponseError.java
@@ -1,4 +1,4 @@
-package com.example.splitly.dto.response;
+package com.example.splitly.presentation.dto.response;
 
 import java.sql.Date;
