+++ src/main/java/com/example/splitly/util/explain.txt
@@ -1,0 +1,1 @@
+Define class helper
